requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
Django==4.1.7
django-appconf==1.0.5
django-extensions==3.2.1
django-filter==23.1
django-select2==8.1.1
Pillow==9.4.0
sqlparse==0.4.3
